refactor(packages): import shared vocabulary from its defining packages
Several package modules imported names from apps/desktop/src/shared/contracts.js
by relative path — a package reaching up into an app. Every such name
originates in a package and was merely re-exported by contracts.ts, so each
import was a hidden package → app → package cycle the declared dependency
graph could not see. Repoint each import at the defining module: the account,
calendar, and superset files at their own packages' modules, and the voice
capability assembler at @sidecar/settings, an edge voice already declares.
A new repository check fails any packages/*/src import whose path contains
apps/, keeping the direction closed.
